add typedoc (#118)
### TL;DR
Added TypeDoc documentation generation to the SDK. 

NOTE: For now, it just works locally, but I intend on adding a github action to deploy them as a public-facing doc in the future (not now)

### What changed?
- Added TypeDoc as a dev dependency
- Created a new `sdk:docs` script that generates documentation
- Added `/docs/` to `.gitignore`
- Configured TypeDoc to exclude readme from generated docs
- Added `.nojekyll` file creation to prevent GitHub Pages processing

### How to test?
1. Install dependencies with `npm install`
2. Run `npm run sdk:docs`
3. Verify documentation is generated in the `/docs` directory
4. Open the index.html file from `/docs`

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/S215IBu9kfHXNs3PdXXg/f058ef04-6562-4666-8493-90e0533fc3e2.png)


### Why make this change?
To provide automatically generated API documentation for the SDK, making it easier for developers to understand and use the available functionality. The TypeDoc integration creates structured documentation from TypeScript code comments and type definitions.