chore: testing - add first testing as an example (#120)
Added 
* https://modern-web.dev/guides/test-runner/getting-started/
* https://www.chaijs.com/ 
* https://playwright.dev/
* https://sinonjs.org/

you can test any of the new npm commands at the root level 
```
    "sdk:test": "web-test-runner \"src/**/*.test.ts\" --node-resolve",
    "sdk:test:coverage": "npm run sdk:test -- --coverage",
    "sdk:test:multiBrowsers": "npm run sdk:test -- --playwright --browsers chromium firefox webkit",
    "sdk:test:watch": "npm run sdk:test -- --watch",
```

Note: I added a `tsconfig.test.json` file just for the IDE, so it won't complain while working on ts tests locally. This is not really used for any actual compilation, as it only processes tests files and NOT production files. This should be a comment in the file itself, but... JSON doesn't support comments 😓